refactor(local-ci): split test_local_ci.py per source module (P11-3)
Roadmap item P11-3. The R2-1 refactor extracted local_ci.py into focused
modules (state_paths, normalize, io_utils, footprint, job_queue, targets,
github_workflows, provenance, ...). This mirrors that on the test side by
relocating the test methods whose subject is wholly an extracted module
out of the 8,584-line test_local_ci.py monolith into per-module files.

Pure mechanical relocation — every moved test method body is byte-verbatim
(same name, same assertions). 15 tests moved into 7 new files; the
remaining 203 tests (whose subject still lives in local_ci.py itself, or
which span multiple modules) stay in test_local_ci.py. Mixed-concern tests
were deliberately left in place rather than force-fit.

New files (own load_module() + class + verbatim shared setUp/tearDown,
matching the established test_local_ci_extra.py pattern):
  test_state_paths.py       (131 lines, 1 test)
  test_io_utils.py          (163 lines, 2 tests)
  test_footprint.py         (149 lines, 1 test)
  test_job_queue.py         (163 lines, 1 test)
  test_targets.py           (146 lines, 1 test)
  test_provenance.py        (150 lines, 1 test)
  test_github_workflows.py  (396 lines, 8 tests)

test_local_ci.py: 8,584 -> 8,183 lines (-401).

Baseline 218 tests passing; post-split 203 + 15 = 218, unchanged.
CI discovery is glob-based (tools/local-ci/test_*.py in
run_python_coverage.py, which runs each file as its own subprocess) so
the new files are picked up with no explicit-list update needed.

Skill-Update: skip skill=ci reason="mechanical test-file relocation — test methods moved verbatim from test_local_ci.py into per-module siblings; no CI workflow, runner, or local-ci behavior change"
