+++ tools/local-ci/test_local_ci.py
@@ -297,45 +297,6 @@ class LocalCiTests(unittest.TestCase):
             else:
                 os.environ["PULP_LOCAL_CI_CONFIG"] = original_override
 
-    def test_size_helpers_report_files_and_local_ci_state_footprint(self):
-        self.assertEqual(self.mod.format_size_bytes(None), "")
-        self.assertEqual(self.mod.format_size_bytes(""), "")
-        self.assertEqual(self.mod.format_size_bytes(512), "512 B")
-        self.assertEqual(self.mod.format_size_bytes(1536), "1.5 KB")
-        self.assertEqual(self.mod.path_size_bytes(self.state_dir / "missing"), 0)
-
-        bundle_dir = self.state_dir / "bundles"
-        bundle_dir.mkdir(parents=True)
-        (bundle_dir / "a.bundle").write_bytes(b"abc")
-        nested = bundle_dir / "nested"
-        nested.mkdir()
-        (nested / "b.bundle").write_bytes(b"defg")
-
-        self.assertEqual(self.mod.path_size_bytes(bundle_dir), 7)
-        footprint = self.mod.local_ci_state_footprint()
-        self.assertEqual(footprint["entries"]["bundles"]["size_bytes"], 7)
-        self.assertEqual(footprint["total_bytes"], 7)
-        self.assertEqual(self.mod.describe_path_for_cleanup(bundle_dir), "bundles")
-        outside = Path(self.tmpdir.name) / "outside"
-        self.assertEqual(self.mod.describe_path_for_cleanup(outside), str(outside))
-
-    def test_text_file_helpers_tail_trim_and_atomic_write(self):
-        log_path = self.state_dir / "logs" / "job" / "mac.log"
-        log_path.parent.mkdir(parents=True)
-        log_path.write_text("one\ntwo\nthree\n")
-
-        self.assertEqual(self.mod.tail_lines(self.state_dir / "missing.log"), [])
-        self.assertEqual(self.mod.tail_lines(log_path, limit=2), ["two\n", "three\n"])
-
-        target_path = self.state_dir / "results" / "result.txt"
-        self.mod.atomic_write_text(target_path, "first\n")
-        self.mod.atomic_write_text(target_path, "second\n")
-        self.assertEqual(target_path.read_text(), "second\n")
-
-        trimmed = self.mod.trim_line("  " + ("x" * 170) + "tail  ", max_len=12)
-        self.assertEqual(len(trimmed), 12)
-        self.assertTrue(trimmed.endswith("tail"))
-
     def test_priority_and_target_helpers_normalize_inputs(self):
         self.assertIsNone(self.mod.parse_targets_arg(""))
         self.assertEqual(
@@ -357,209 +318,6 @@ class LocalCiTests(unittest.TestCase):
         }
         self.assertEqual(self.mod.resolve_targets(config, None), ["mac", "ubuntu"])
         self.assertEqual(self.mod.resolve_targets(config, ["windows", "mac", "windows"]), ["mac", "windows"])
-
-    def test_extracted_target_helpers_cover_empty_defaults_unknown_and_disabled(self):
-        config = {
-            "targets": {
-                "mac": {"enabled": True},
-                "linux": {"enabled": False},
-                "windows": {},
-            },
-            "defaults": {"targets": []},
-        }
-
-        self.assertEqual(self.mod.enabled_targets(config), ["mac", "windows"])
-        self.assertEqual(self.mod.resolve_targets(config, None), [])
-
-        with self.assertRaisesRegex(ValueError, "Unknown target\\(s\\): ios"):
-            self.mod.resolve_targets(config, ["ios"])
-
-        with self.assertRaisesRegex(ValueError, "Requested target\\(s\\) disabled"):
-            self.mod.resolve_targets(config, ["linux"])
-
-    def test_extracted_provenance_helpers_normalize_result_and_summarize(self):
-        direct = self.mod.normalize_provenance({"direct_backend": ""})
-        self.assertEqual(direct["execution_kind"], "direct")
-        self.assertEqual(direct["direct_backend"], "")
-        self.assertEqual(self.mod.provenance_summary(direct), "direct via local-ci")
-
-        hosted = self.mod.normalize_provenance(
-            {
-                "execution_kind": "hosted",
-                "hosted_orchestrator": "github-actions",
-                "runner_provider": "github-hosted",
-                "runner_selector": "macos-15",
-                "run_id": "12345",
-            }
-        )
-        self.assertEqual(
-            self.mod.provenance_summary(hosted),
-            "hosted via github-actions/github-hosted selector=macos-15 run=12345",
-        )
-
-        result = self.mod.normalize_result({"submission": {"provenance": hosted}})
-        self.assertEqual(result["provenance"]["run_id"], "12345")
-
-    def test_extracted_queue_helpers_accept_legacy_array_and_enveloped_queue(self):
-        legacy_job = {
-            "branch": "feature/local-ci",
-            "sha": "abc123",
-            "queued_at": "2026-05-18T00:00:00Z",
-            "priority": " HIGH ",
-            "targets": ["windows", "mac", "mac"],
-            "validation": " smoke ",
-            "submission": {
-                "provenance": {
-                    "execution_kind": "hosted",
-                    "hosted_orchestrator": "github-actions",
-                    "runner_provider": "github-hosted",
-                }
-            },
-        }
-
-        normalized = self.mod.normalize_job(legacy_job)
-        self.assertEqual(len(normalized["id"]), 12)
-        self.assertEqual(normalized["priority"], "high")
-        self.assertEqual(normalized["targets"], ["mac", "windows"])
-        self.assertEqual(normalized["validation"], "smoke")
-        self.assertEqual(normalized["status"], "pending")
-        self.assertEqual(normalized["provenance"]["execution_kind"], "hosted")
-
-        queue_file = self.mod.queue_path()
-        queue_file.parent.mkdir(parents=True, exist_ok=True)
-        queue_file.write_text(json.dumps([legacy_job]) + "\n")
-        self.assertEqual(self.mod.load_queue_unlocked()[0]["targets"], ["mac", "windows"])
-
-        queue_file.write_text(json.dumps({"jobs": [legacy_job]}) + "\n")
-        self.assertEqual(self.mod.load_queue_unlocked()[0]["priority"], "high")
-
-        self.mod.save_queue_unlocked([normalized])
-        self.assertEqual(json.loads(queue_file.read_text())[0]["id"], normalized["id"])
-
-    def test_extracted_github_workflow_helpers_resolve_sources_and_cli_overrides(self):
-        config = {
-            "github_actions": {
-                "defaults": {"provider": "namespace"},
-                "workflows": {
-                    "build": {
-                        "providers": {
-                            "namespace": {
-                                "linux_runner_selector_json": '["ns-linux"]',
-                            }
-                        }
-                    }
-                },
-            }
-        }
-        repo_vars = {
-            "PULP_NAMESPACE_BUILD_WINDOWS_RUNS_ON_JSON": '"ns-windows"',
-        }
-
-        defaults, sources = self.mod.resolve_workflow_dispatch_defaults(
-            config,
-            repo_vars,
-            "build",
-            "namespace",
-            ["linux_runner_selector_json", "windows_runner_selector_json"],
-        )
-        self.assertEqual(defaults["linux_runner_selector_json"], '["ns-linux"]')
-        self.assertEqual(defaults["windows_runner_selector_json"], '"ns-windows"')
-        self.assertIn("config github_actions.workflows.build", sources["linux_runner_selector_json"])
-        self.assertIn("repo variable", sources["windows_runner_selector_json"])
-
-        args = SimpleNamespace(
-            linux_runner_selector_json='"ubuntu-24.04"',
-            windows_runner_selector_json=None,
-            macos_runner_selector_json=None,
-        )
-        self.assertEqual(
-            self.mod.resolve_cli_dispatch_field_values(args, ["linux_runner_selector_json"]),
-            {"linux_runner_selector_json": '"ubuntu-24.04"'},
-        )
-
-        with self.assertRaisesRegex(ValueError, "not supported for this workflow"):
-            self.mod.resolve_cli_dispatch_field_values(args, ["macos_runner_selector_json"])
-
-        with self.assertRaisesRegex(ValueError, "Unknown workflow"):
-            self.mod.resolve_default_provider_for_workflow(
-                {"provider": "github-hosted"}, "missing"
-            )
-
-    def test_extracted_github_workflow_settings_validate_selectors_and_providers(self):
-        config = {
-            "github_actions": {
-                "repository": "  danielraffel/pulp  ",
-                "defaults": {
-                    "workflow": "  docs-check  ",
-                    "provider": "  namespace  ",
-                    "wait_poll_secs": "11",
-                    "match_timeout_secs": 22,
-                },
-                "workflows": {
-                    "docs-check": {
-                        "providers": {
-                            "namespace": {
-                                "runner_selector_json": '["namespace-profile-docs"]',
-                            }
-                        }
-                    }
-                },
-            }
-        }
-
-        display = self.mod.github_actions_settings_for_display(config)
-        self.assertEqual(display["repository"], "danielraffel/pulp")
-        self.assertEqual(display["workflow"], "docs-check")
-        self.assertEqual(display["provider"], "namespace")
-
-        settings = self.mod.resolve_github_actions_settings(config)
-        self.assertEqual(settings["wait_poll_secs"], 11)
-        self.assertEqual(settings["match_timeout_secs"], 22)
-
-        selector = self.mod.resolve_workflow_runner_selector_json(
-            config, "docs-check", "namespace"
-        )
-        self.assertEqual(selector, '["namespace-profile-docs"]')
-        self.assertEqual(
-            self.mod.resolve_workflow_runner_selector_json(
-                {"github_actions": {"workflows": []}}, "docs-check", "namespace"
-            ),
-            "",
-        )
-
-        self.assertEqual(
-            self.mod.normalize_runs_on_json('"macos-15"', setting_name="selector"),
-            '"macos-15"',
-        )
-        self.assertEqual(
-            self.mod.normalize_runs_on_json('["ubuntu-24.04"]', setting_name="selector"),
-            '["ubuntu-24.04"]',
-        )
-        with self.assertRaisesRegex(ValueError, "must be valid JSON"):
-            self.mod.normalize_runs_on_json("not json", setting_name="selector")
-        with self.assertRaisesRegex(ValueError, "must decode to a string or array"):
-            self.mod.normalize_runs_on_json("42", setting_name="selector")
-
-        with self.assertRaisesRegex(ValueError, "must be positive"):
-            self.mod.resolve_github_actions_settings(
-                {"github_actions": {"defaults": {"wait_poll_secs": 0}}}
-            )
-
-        provider, source = self.mod.resolve_default_provider_for_workflow(
-            {"provider": "namespace"}, "validate"
-        )
-        self.assertEqual(provider, "github-hosted")
-        self.assertIn("workflow fallback", source)
-
-        provider, source = self.mod.resolve_default_provider_for_workflow(
-            {"provider": "github-hosted"}, "build", explicit_provider="namespace"
-        )
-        self.assertEqual((provider, source), ("namespace", "cli"))
-
-        with self.assertRaisesRegex(ValueError, "does not support provider"):
-            self.mod.resolve_default_provider_for_workflow(
-                {"provider": "github-hosted"}, "validate", explicit_provider="namespace"
-            )
 
     def test_windows_checkout_path_helpers_join_and_detect_unsafe_roots(self):
         self.assertEqual(
@@ -585,24 +343,6 @@ class LocalCiTests(unittest.TestCase):
                 r"C:\Users\daniel",
             )
         )
-
-    def test_image_change_summary_falls_back_to_hash_for_non_images(self):
-        before_path = Path(self.tmpdir.name) / "before.bin"
-        after_path = Path(self.tmpdir.name) / "after.bin"
-        diff_path = Path(self.tmpdir.name) / "diff" / "image.png"
-        before_path.write_bytes(b"same")
-        after_path.write_bytes(b"same")
-
-        unchanged = self.mod.image_change_summary(before_path, after_path, diff_output_path=diff_path)
-        self.assertFalse(unchanged["changed"])
-        self.assertEqual(unchanged["method"], "file-hash")
-        self.assertFalse(diff_path.exists())
-
-        after_path.write_bytes(b"different")
-        changed = self.mod.image_change_summary(before_path, after_path, diff_output_path=diff_path)
-        self.assertTrue(changed["changed"])
-        self.assertEqual(changed["method"], "file-hash")
-        self.assertFalse(diff_path.exists())
 
     def test_resolve_submission_options_uses_branch_tip_when_branch_is_explicit(self):
         args = SimpleNamespace(
@@ -6143,10 +5883,6 @@ class LocalCiTests(unittest.TestCase):
         self.assertIn("prefix", logged)
         self.assertIn("suffix", logged)
 
-    def test_target_log_path_uses_machine_global_logs_dir(self):
-        path = self.mod.target_log_path("job123", "windows")
-        self.assertEqual(path, self.state_dir / "logs" / "job123" / "windows.log")
-
     def test_save_result_updates_evidence_index_with_last_good_target_results(self):
         result_path_one = self.mod.save_result(
             {
@@ -6449,58 +6185,6 @@ class LocalCiTests(unittest.TestCase):
 
         self.assertIn("Execution: `direct via local-ci`", comment)
 
-    def test_resolve_github_actions_settings_reads_optional_config_defaults(self):
-        settings = self.mod.resolve_github_actions_settings(self.mod.load_optional_config())
-        self.assertEqual(settings["repository"], "danielraffel/pulp")
-        self.assertEqual(settings["workflow"], "build")
-        self.assertEqual(settings["provider"], "github-hosted")
-        self.assertEqual(settings["wait_poll_secs"], 5)
-        self.assertEqual(settings["match_timeout_secs"], 30)
-
-    def test_resolve_workflow_runner_selector_json_reads_docs_check_provider_default(self):
-        selector = self.mod.resolve_workflow_runner_selector_json(
-            self.mod.load_optional_config(), "docs-check", "namespace"
-        )
-        self.assertEqual(selector, "\"namespace-profile-default\"")
-
-    def test_resolve_default_provider_for_workflow_falls_back_to_github_hosted(self):
-        provider, source = self.mod.resolve_default_provider_for_workflow(
-            {"provider": "namespace"},
-            "validate",
-        )
-        self.assertEqual(provider, "github-hosted")
-        self.assertIn("workflow fallback", source)
-
-    def test_resolve_workflow_field_value_and_source_reads_repo_variable_fallback(self):
-        config = json.loads(self.config_path.read_text())
-        del config["github_actions"]["workflows"]["docs-check"]["providers"]["namespace"]["runner_selector_json"]
-        self.config_path.write_text(json.dumps(config) + "\n")
-
-        value, source = self.mod.resolve_workflow_field_value_and_source(
-            self.mod.load_optional_config(),
-            {"PULP_NAMESPACE_DOCS_CHECK_RUNS_ON_JSON": "\"namespace-profile-repo-var\""},
-            "docs-check",
-            "namespace",
-            "runner_selector_json",
-        )
-        self.assertEqual(value, "\"namespace-profile-repo-var\"")
-        self.assertEqual(source, "repo variable PULP_NAMESPACE_DOCS_CHECK_RUNS_ON_JSON")
-
-    def test_resolve_workflow_dispatch_field_values_reads_build_namespace_defaults(self):
-        fields = self.mod.resolve_workflow_dispatch_field_values(
-            self.mod.load_optional_config(),
-            "build",
-            "namespace",
-            ["linux_runner_selector_json", "windows_runner_selector_json"],
-        )
-        self.assertEqual(
-            fields,
-            {
-                "linux_runner_selector_json": "\"namespace-profile-default\"",
-                "windows_runner_selector_json": "\"namespace-profile-default\"",
-            },
-        )
-
     def test_cloud_record_round_trip_and_lookup(self):
         first = self.mod.normalize_cloud_record(
             {
@@ -6703,91 +6387,6 @@ class LocalCiTests(unittest.TestCase):
             self.mod.normalize_runs_on_json("ubuntu-latest", setting_name="runs-on")
         with self.assertRaisesRegex(ValueError, "string or array"):
             self.mod.normalize_runs_on_json('{"label": "ubuntu"}', setting_name="runs-on")
-
-    def test_workflow_selector_helpers_resolve_config_repo_and_cli_values(self):
-        config = {
-            "github_actions": {
-                "defaults": {"provider": "namespace"},
-                "workflows": {
-                    "build": {
-                        "providers": {
-                            "namespace": {
-                                "linux_runner_selector_json": '"linux-config"',
-                            }
-                        }
-                    },
-                    "docs-check": {
-                        "providers": {
-                            "namespace": {},
-                        }
-                    },
-                },
-            }
-        }
-        repo_variables = {
-            "PULP_NAMESPACE_BUILD_WINDOWS_RUNS_ON_JSON": '"windows-repo"',
-            "PULP_NAMESPACE_BUILD_MACOS_RUNS_ON_JSON": '["self-hosted", "macOS"]',
-            "PULP_NAMESPACE_DOCS_CHECK_RUNS_ON_JSON": '"docs-repo"',
-        }
-        settings = self.mod.resolve_github_actions_settings(config)
-
-        build_summary = self.mod.summarize_workflow_provider_defaults(
-            config, repo_variables, settings, "build"
-        )
-        self.assertEqual(build_summary["provider"], "namespace")
-        self.assertEqual(
-            build_summary["dispatch_fields"],
-            {
-                "linux_runner_selector_json": '"linux-config"',
-                "windows_runner_selector_json": '"windows-repo"',
-                "macos_runner_selector_json": '["self-hosted", "macOS"]',
-            },
-        )
-        self.assertEqual(
-            build_summary["dispatch_sources"]["linux_runner_selector_json"],
-            "config github_actions.workflows.build.providers.namespace.linux_runner_selector_json",
-        )
-        self.assertEqual(
-            build_summary["dispatch_sources"]["windows_runner_selector_json"],
-            "repo variable PULP_NAMESPACE_BUILD_WINDOWS_RUNS_ON_JSON",
-        )
-
-        docs_summary = self.mod.summarize_workflow_provider_defaults(
-            config, repo_variables, settings, "docs-check"
-        )
-        self.assertEqual(docs_summary["selector_value"], '"docs-repo"')
-        self.assertEqual(
-            docs_summary["selector_source"],
-            "repo variable PULP_NAMESPACE_DOCS_CHECK_RUNS_ON_JSON",
-        )
-
-        cli_fields = self.mod.resolve_cli_dispatch_field_values(
-            SimpleNamespace(
-                linux_runner_selector_json='"linux-cli"',
-                windows_runner_selector_json=None,
-                macos_runner_selector_json=None,
-            ),
-            ["linux_runner_selector_json"],
-        )
-        self.assertEqual(cli_fields, {"linux_runner_selector_json": '"linux-cli"'})
-        with self.assertRaisesRegex(ValueError, "not supported"):
-            self.mod.resolve_cli_dispatch_field_values(
-                SimpleNamespace(
-                    linux_runner_selector_json='"linux-cli"',
-                    windows_runner_selector_json=None,
-                    macos_runner_selector_json=None,
-                ),
-                [],
-            )
-        with self.assertRaisesRegex(ValueError, "valid JSON"):
-            self.mod.resolve_cli_dispatch_field_values(
-                SimpleNamespace(
-                    linux_runner_selector_json="linux-cli",
-                    windows_runner_selector_json=None,
-                    macos_runner_selector_json=None,
-                ),
-                ["linux_runner_selector_json"],
-            )
 
     def test_cloud_record_lookup_and_summary_helpers_cover_edge_statuses(self):
         normalized = self.mod.normalize_cloud_record(
